+++ client/src/shared/ui/date/Date.module.scss
@@ -20,10 +20,6 @@
 		transition: all;
 		transition-duration: 300ms;
 
-		&_textarea {
-			resize: none !important;
-		}
-
 		&_inputErr {
 			outline: 1px solid #e03737;
 		}
