+++ client/src/widgets/ClientInfo/ui/ClientInfo.module.scss
@@ -39,23 +39,16 @@
 		}
 
 		&_infoBlock {
-			flex-wrap: wrap;
-			justify-content: space-between;
-			display: flex;
+			display: grid;
+			grid-template-columns: 1fr 1fr 1fr;
+			gap: 0 30px;
 			padding: 18px;
 			background-color: #d4d4d4;
 			border-radius: 20px;
+			z-index: 2;
 
 			&_info {
-				z-index: 3;
-				display: flex;
-				flex-direction: column;
-				gap: 8px;
-
-				&_item {
-					max-width: 400px;
-					width: 400px;
-				}
+				z-index: 2;
 			}
 		}
 	}
