+++ client/src/widgets/ClientInfo/ui/ClientInfo.tsx
@@ -8,7 +8,7 @@ import { EditFilled } from '@ant-design/icons'
 import { DatePicker, Input, Select } from '@shared/ui'
 
 import cl from './ClientInfo.module.scss'
-import { firstClientInfo, secondClientInfo, thirdClientInfo } from './data'
+import { clientInfo } from './clietnInfo.data'
 
 const ClientInfo = () => {
 	type TClientInfo = {
@@ -84,159 +84,60 @@ const ClientInfo = () => {
 					</button>
 				</div>
 				<div className={cl.root__container_infoBlock}>
-					<div className={cl.root__container_infoBlock_info}>
-						{firstClientInfo.map(({ isDatepicker, rules, ...i }) =>
-							i.options ? (
-								<Controller
-									name={i.name}
-									control={control}
-									key={i.name}
-									render={({ field }) => (
-										<Select
+					{clientInfo.map(({ isDatepicker, rules, ...i }) =>
+						i.options ? (
+							<Controller
+								name={i.name}
+								control={control}
+								key={i.name}
+								render={({ field }) => (
+									<Select
+										disabled={isDisabled}
+										field={field}
+										placeholder={i.name}
+										bodyClassName={
+											cl.root__container_infoBlock_info
+										}
+										{...i}
+									/>
+								)}
+							/>
+						) : (
+							<Controller
+								name={i.name}
+								control={control}
+								key={i.name}
+								rules={rules}
+								render={({ field }) =>
+									isDatepicker ? (
+										<DatePicker
+											style={{
+												zIndex: 2
+											}}
 											disabled={isDisabled}
 											field={field}
-											placeholder={i.name}
-											bodyClassName={
-												cl.root__container_infoBlock_info_item
+											className={
+												cl.root__container_infoBlock_info
 											}
+											format='DD-MM-YY'
+											error={errors[i.name]?.message}
 											{...i}
 										/>
-									)}
-								/>
-							) : (
-								<Controller
-									name={i.name}
-									control={control}
-									key={i.name}
-									rules={rules}
-									render={({ field }) =>
-										isDatepicker ? (
-											<DatePicker
-												// common={
-												// 	field.value
-												// 		? `*${new Date().getFullYear() - new Date(field.value).getFullYear()} лет`
-												// 		: ''
-												// }
-												disabled={isDisabled}
-												field={field}
-												style={{ maxHeight: '45.14px' }}
-												className={
-													cl.root__container_infoBlock_info_item
-												}
-												format='DD-MM-YY'
-												error={errors[i.name]?.message}
-												{...i}
-											/>
-										) : (
-											<Input
-												disabled={isDisabled}
-												field={field}
-												bodyClassName={
-													cl.root__container_infoBlock_info_item
-												}
-												error={errors[i.name]?.message}
-												{...i}
-											/>
-										)
-									}
-								/>
-							)
-						)}
-					</div>
-					<div className={cl.root__container_infoBlock_info}>
-						{secondClientInfo.map(({ rules, ...i }) =>
-							i.options ? (
-								<Controller
-									name={i.name}
-									control={control}
-									key={i.name}
-									render={({ field }) => (
-										<Select
-											disabled={isDisabled}
-											field={field}
-											placeholder={i.name}
-											bodyClassName={
-												cl.root__container_infoBlock_info_item
-											}
-											{...i}
-										/>
-									)}
-								/>
-							) : (
-								<Controller
-									name={i.name}
-									control={control}
-									key={i.name}
-									rules={rules}
-									render={({ field }) => (
+									) : (
 										<Input
 											disabled={isDisabled}
 											field={field}
 											bodyClassName={
-												cl.root__container_infoBlock_info_item
+												cl.root__container_infoBlock_info
 											}
 											error={errors[i.name]?.message}
 											{...i}
 										/>
-									)}
-								/>
-							)
-						)}
-					</div>
-					<div className={cl.root__container_infoBlock_info}>
-						{thirdClientInfo.map(({ isDatepicker, rules, ...i }) =>
-							i.options ? (
-								<Controller
-									name={i.name}
-									control={control}
-									key={i.name}
-									render={({ field }) => (
-										<Select
-											disabled={isDisabled}
-											field={field}
-											placeholder={i.name}
-											bodyClassName={
-												cl.root__container_infoBlock_info_item
-											}
-											{...i}
-										/>
-									)}
-								/>
-							) : (
-								<Controller
-									name={i.name}
-									control={control}
-									key={i.name}
-									rules={rules}
-									render={({ field }) =>
-										isDatepicker ? (
-											<DatePicker
-												disabled={isDisabled}
-												field={field}
-												style={{ maxHeight: '45.14px' }}
-												className={
-													cl.root__container_infoBlock_info_item
-												}
-												format='DD-MM-YY'
-												error={errors[i.name]?.message}
-												{...i}
-											/>
-										) : (
-											<Input
-												disabled={isDisabled}
-												field={field}
-												bodyClassName={
-													cl.root__container_infoBlock_info_item
-												}
-												error={errors[i.name]?.message}
-												{...i}
-											/>
-										)
-									}
-								/>
-							)
-						)}
-					</div>
+									)
+								}
+							/>
+						)
+					)}
 				</div>
 			</div>
 		</form>
