--- client/src/widgets/ClientInfo/ui/data/index.ts
@@ -1,3 +1,0 @@
-export * from './firstClientsInfo.data'
-export * from './secondClientsInfo.data'
-export * from './thirdClientsInfo.data'
